refactor(api): move test auth bypass to test infrastructure
Address code review comments:
- Remove test environment authentication bypass from Program.cs
- Delete TestAuthorizationHandler.cs from main API (duplicate in test assembly)
- Move HeaderBasedCurrentUserService.cs to test infrastructure
- Remove unused using statements from Program.cs
- Update XML comment to reference MediaTypeNames.Application.Json

Test infrastructure should configure its own services via WebApplicationFactory,
not rely on environment-specific code in the main application.
